MacOS: New updater model
It turned out that launchd(8) does not like it when you're a daemon
running as root and you fork a child which drops privileges.  What
happens, in particular, is that the system starts /sbin/launchd for
the user you dropped privileges to -- eventually the extra copy of
launchd(8) exit()s, but this isn't relevant: the problem here's that
this process is indeed created.  I spent quite some time wrestling
with the problem, to no avail and finally I decided to break the
updater into pieces, using launchd(8).

This patch splits the updater into two pieces: the privileged "core"
and the unprivileged "dload".  And uses launchd(8) to start, stop
and control them.  The "core" updater (org.neubot.updater.core) is a
job configured to run every 30 minutes, with full root privileges,
to check for updates and eventually update Neubot.  As in the
previous model, it does not communicate directly with the network.
This task is performed by "dload", an unprivileged job, which is
started on demand by launchd(8) (org.neubot.updater.dload).

The two jobs communicate using ``/var/run/neubot-dload.sock``, which
is a UNIX domain socket.  Whenever it wants to fetch a file, the
core job connects to the UNIX domain socket.  In turn, launchd(8)
accept()s the connection and starts the dload job, redirecting its
stdio descriptors to the connected socket.  So the dload job will
read the URI to download from its standard input and will write the
downloaded data to its standard output.

The communication protocol is very simple.  The request is a line
that contains just the URI to download.  And the response consists
of an header line followed by the downloaded data.  The header line
contains the word "OK" followed by the data lenght.  If something
goes wrong, the dload job will just exit() and the core job will
not read anything from the socket.

The core job logic is a copy and paste of part of the original
updater logic, with the difference that it does not take care of
the agent process.  Instead, the core job assumes it can use
launchctl(1) to stop the agent when a new version is available.

This patch also updates the execution model.  Now the Neubot agent
is started by launchctl(8) and, since the beginning, it runs on behalf
of the unprivileged user ``_neubot``.

While there, also fix a comment in ``makepkg.py``.