WORK IN PROGRESS Clean stop and restart
To implement a clean stop and restart I need basically to rely
again on launchctl(1).  This is not something that can be performed
in a robust way by the updater job.

This is the major showstopped for this piece of code to enter
the real testing phase.  I'm sorry but this means that will not
be merged in the upcoming 0.4.3.