regress/Debian/control/prerm: repair viewer/notifier test
Now the viewer and the notifier do not fail anymore if the database
does not exist, because they use the default Neubot datatabase, which
main() tries to open.  Still, after this change, this regress fails
because the Neubot user does not exist.  So, this commit moves the
code to create Neubot user and group before the code that starts the
viewer and the notifier.  Additionally, it gives them more time to
start, since the viewer takes about 15 seconds to decide that Neubot
is not running and it makes sense to show its GUI.  Finally zap the
/var/neubot directory, which is created by main() when it tries to
open the default database.